ScriptableObject settings have been added to almost all parts of the AISystem
AIManager has AIManagerSettings.
This is where the user can determine what types of enemies are contained within each object pool. These will be the enemy types that the spawners will be able to use.

SpawnSettings
This is a single point of control that the spawners will be able to use. If several spawners require the same settings then it will be easier to add.