WIP Spawning will be based on players vision
Currently checking an AABB of the agent size using the camera frustum. If it's on screen it is not a valid spawn location.
TO DO: Check if there are objects obsuring the vision of these location.
Fix bug where if there is no valid spawn the agents will spawn at origin.