Spawning can be based on the players vision of the spawn Location
This is first checked if the bounds of the agent would be seen on camera and then raycasts to each corner of the bounds to check if there are obstacles
ToDo:
Agents still spawn at vector3.zero when no valid location is found.
Add random modifier values when determining which location to spawn in. (These could be things like, how often should enemies spawn behind the player, how often should they spawn close vs far away to the player.)