Further fleshed out Room Placer Tool
It can be toggle so that a Gameobject will follow another gameobject but that second game object won't move if the first does. For example, an lamp on a desk (object1) will follow a room (object2) that is moved. If the lamp moves the room will stay where it is.